requirements.txt
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
langsmith>=0.1.0
python-dotenv>=1.0.0
pydantic>=2.9.0
httpx>=0.27.0
